+++ packages/terra-status/src/terra-dev-site/test/status/StatusNoColor.test.jsx
@@ -1,7 +1,12 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Status from '../../../Status';
+import styles from '../../doc/example/colors.module.scss';
 
-const simpleText = <div style={{ padding: '5px' }}>Sample text</div>;
+const cx = classNames.bind(styles);
+
+
+const simpleText = <div className={cx('text-wrapper')}>Sample text</div>;
 
 const StatusNoColor = () => (
   <div>